feat(board-factory): 다양한 Board 생성 역할을 담당할 BoardFactory 추가
